[FLINK-9377] [core] (part 3) Deprecate TypeSerializerSerializationUtil
This commit deprecates all utility methods and classes related to
serializing serializers. All methods that will still be in use, i.e.
writing config snapshots, are now moved to a separate new
TypeSerializerConfigSnapshotSerializationUtil class.
